some polishing, moving more to internal
